refactor: move FIM prompt building to HoleFiller class
- Add getFimPrompts method to HoleFiller for FIM-specific prompt building
- Parallel structure to getPrompts for chat-based completions
- Remove FIM prompt building logic from GhostInlineCompletionProvider.getPrompt
- Centralizes all prompt building logic in HoleFiller class
- All tests pass (63/63)
